<Vitalii Koshura>: Improve HoG detector. Make drawing HoG detectors

+++ src/libpose/surfDetector.cpp
@@ -502,7 +502,7 @@ LimbLabel SurfDetector::generateLabel(Frame *frame, BodyPart bodyPart, Point2f j
   maskMat.release();
   imgMat.release();
   float score = compare(bodyPart, partModel, j0, j1);
-  score += inMaskPixels / totalPixels;
+  score += ((float)inMaskPixels / (float)totalPixels);
   Score sc(score, detectorName.str(), _useSURFdet);
   s.push_back(sc);
   return LimbLabel(bodyPart.getPartID(), boxCenter, rot, rect.asVector(), s);
